+++ DoorDeliveryApplication/src/UI/Customer/ViewCustomerJPanel.java
@@ -217,7 +217,7 @@ public class ViewCustomerJPanel extends javax.swing.JPanel {
         }
         
         String name = txtName.getText();
-        long phone = Long.parseLong(txtPhone.getText());
+        int phone = Integer.parseInt(txtPhone.getText());
         int zipcode = Integer.parseInt(txtZipcode.getText());
         String streetaddress = txtAddress.getText();
         String emailaddress = txtEmail.getText();
@@ -249,6 +249,7 @@ public class ViewCustomerJPanel extends javax.swing.JPanel {
         sysAdminwjp.populateTable();
         CardLayout layout = (CardLayout) userProcessContainer.getLayout();
         layout.previous(userProcessContainer);
+        
     }//GEN-LAST:event_btnBackActionPerformed
 
     private void btnUpdateActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnUpdateActionPerformed
@@ -262,6 +263,7 @@ public class ViewCustomerJPanel extends javax.swing.JPanel {
         txtPhone.setEditable(true);
         txtZipcode.setEditable(true);
         btnSave.setEnabled(true);
+        
     }//GEN-LAST:event_btnUpdateActionPerformed
